+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Select: fix the select dropdown focus not being trapped in it.
+
 ## [3.2.0][] - 2023-04-11
 
 ### Changed